requirements.txt
fastapi
uvicorn
jinja2
pgconnect
colorama
pytz
python-dotenv
google-auth
requests
python-multipart
passlib
cachetools
websockets
aiohttp
aiosmtplib
python-dateutil
aiocache
psutil
discord